Created MotifTile
